+++ CHANGELOG.md
@@ -9,6 +9,7 @@ master
 * Log with Growl on 10.7, with notification center on 10.8 and up. Fix #59.
 * Use windowDidLoad rather than override showWindow. Fix #62.
 * Never die while processing a screenshot, log exceptions as needed. Fix #38.
+* Parse timestamp out of screenshot filename to not reprocess old files. Fix #40.
 
 v2.0 (8/22/2014)
 ----------------